Allow setting combobox debounce value, make combobox controlled/uncontrolled mutually exclusive
